fix: allow image input to be optional
When creating a cupcake, the image URL input is supposed to be
optional.  To allow for that, retrieving the value from request.json
was changed from a direct access using [] to using the dictionary get
method.  This now will not throw an exception if there is no image key
in the request.json dictionary for the API call, and the Cupcake model
will receive a NoneType and will set the image URL with the default URL.